+++ test/frontend/matcher/completion_test.dart
@@ -16,8 +16,7 @@ void main() {
         expect(10, doesNotComplete);
       });
 
-      expectTestFailed(
-          liveTest,"10 is not a Future");
+      expectTestFailed(liveTest, "10 is not a Future");
     });
 
     test("succeeds when a future does not complete", () {