+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.12.21
 
+* Add a `doesNotComplete` matcher that asserts that a Future never completes.
+
 * `throwsA()` and all related matchers will now match functions that return
   `Future`s that emit exceptions.